code refactoring of component AdverDesktop
