+++ src/components/adver/form/AdverDesktop/AdverDesktop.module.scss
@@ -118,6 +118,7 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  margin-top: 10px;
   width: 335px;
   height: 46px;
   padding: 12px;
