+++ src/components/adver/form/SectionSwitch/SectionSwitch.module.scss
@@ -1,0 +1,45 @@
+.linkItem {
+  display: flex;
+  gap: 20px;
+  align-items: center;
+}
+
+.sellBox {
+  margin-bottom: 40px;
+}
+
+.linkItemText {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  cursor: pointer;
+  color: #131313;
+}
+
+.labelText {
+  font-size: var(--large-body-medium);
+  font-weight: 400;
+}
+
+.radio {
+  appearance: none;
+  width: 32px;
+  height: 32px;
+  border: 1px solid #777676;
+  border-radius: 50%;
+  position: relative;
+  cursor: pointer;
+
+  &:checked::before {
+    content: "";
+    display: block;
+    width: 24px;
+    height: 24px;
+    background-color: #b8dd05;
+    border-radius: 50%;
+    position: absolute;
+    top: 3px;
+    left: 3px;
+  }
+}
+
